DEV2-2972 catch browser creation error gracefully
