+++ Tabnine/src/main/java/com/tabnine/chat/TabnineChatProjectManagerListener.kt
@@ -11,6 +11,7 @@ import com.intellij.openapi.util.IconLoader
 import com.intellij.openapi.wm.RegisterToolWindowTask
 import com.intellij.openapi.wm.ToolWindowAnchor
 import com.intellij.openapi.wm.ToolWindowManager
+import com.tabnine.BrowserCreator
 import com.tabnineCommon.capabilities.CapabilitiesService
 import com.tabnineCommon.capabilities.Capability
 import com.tabnineCommon.lifecycle.BinaryCapabilitiesChangeNotifier
@@ -19,7 +20,7 @@ import javax.swing.JPanel
 
 class TabnineChatProjectManagerListener private constructor() : ProjectManagerListener, Disposable {
     private val initialized = AtomicBoolean(false)
-    private var messageRouter = ChatMessagesRouter()
+    private var messagesRouter = ChatMessagesRouter()
     private val toolWindowIcon = IconLoader.findIcon("/icons/tabnine-tool-window-icon.svg")
 
     companion object {
@@ -71,14 +72,21 @@ class TabnineChatProjectManagerListener private constructor() : ProjectManagerLi
             .info("Chat enabled: $chatEnabled (alpha: $alphaEnabled, chat capability: $chatCapabilityEnabled)")
         if (!chatEnabled) return
 
+        val browser = try {
+            BrowserCreator.createBrowser(messagesRouter, project)
+        } catch (e: Exception) {
+            Logger.getInstance(javaClass).warn("Failed to create browser", e)
+            return
+        }
+
         ToolWindowManager.getInstance(project).registerToolWindow(
             RegisterToolWindowTask(
-                id = "TabnineChat",
+                id = "Tabnine Chat",
                 anchor = ToolWindowAnchor.RIGHT,
                 canCloseContent = false,
                 component = JPanel(),
                 icon = toolWindowIcon,
-                contentFactory = TabnineChatWebViewFactory(messageRouter)
+                contentFactory = TabnineChatWebViewFactory(browser)
             )
         )
     }
